+++ Changes.rst
@@ -1,3 +1,11 @@
+.. currentmodule:: httpio
+
+Unreleased Changes
+==================
+
+* Introduced `BodyFollowing` class for use with *body* parameter of
+  `~HTTPConnection.send_request` method.
+
 Release 1.0 (2013-07-13)
 ========================